Split out constant portion of PhraseSplitter to PhraseSplitterQueryEnv.
